Re-add previous bibliography behaviour, create separate plugin for new behaviour
