+++ LaTeXTools.sublime-settings
@@ -331,8 +331,22 @@
 	},
 
 // ------------------------------------------------------------------
-// Formatting of bibliographic references in quick panel
+// Bibliographic references
 // ------------------------------------------------------------------
+	// OPTION: "bibliography"
+	// Either a single  bibliography plugin to use or a list of plugins 
+	// which will be executed in order, stopping after the first result
+	// found
+	//
+	// Possible values:
+	//
+	// "traditional_bibliography" 		the default, regex-based bibliography
+	//									parsing
+	//
+	// "new_bibliography"				a newer parser which supports UTF-8 and
+	//									additional fields, but may be slower and
+	//									less well tested.
+	"bibliography": "traditional_bibliography",
 
 	/* This preference sets the format of the quick panel to select citations using wildcards.
 	The setting is a list with one or two string using wildcards for author, title, keyword etc.
