+++ bibliography_plugins/traditionalBibliography.py
@@ -1,119 +1,42 @@
 from latextools_plugin import LaTeXToolsPlugin
 
-from bibtex import Parser
-from bibtex.names import Name
-from bibtex.tex import tokenize_list
-
-import latex_chars
 from latextools_utils import cache
 
 import codecs
-from collections import Mapping
 import hashlib
 import os
+import re
 import sublime
 import traceback
 
-# LaTeX -> Unicode decoder
-latex_chars.register()
+kp = re.compile(r'@[^\{]+\{(.+),')
+# new and improved regex
+# we must have "title" then "=", possibly with spaces
+# then either {, maybe repeated twice, or "
+# then spaces and finally the title
+# # We capture till the end of the line as maybe entry is broken over several lines
+# # and in the end we MAY but need not have }'s and "s
+# tp = re.compile(r'\btitle\s*=\s*(?:\{+|")\s*(.+)', re.IGNORECASE)  # note no comma!
+# # Tentatively do the same for author
+# # Note: match ending } or " (surely safe for author names!)
+# ap = re.compile(r'\bauthor\s*=\s*(?:\{|")\s*(.+)(?:\}|"),?', re.IGNORECASE)
+# # Editors
+# ep = re.compile(r'\beditor\s*=\s*(?:\{|")\s*(.+)(?:\}|"),?', re.IGNORECASE)
+# # kp2 = re.compile(r'([^\t]+)\t*')
+# # and year...
+# # Note: year can be provided without quotes or braces (yes, I know...)
+# yp = re.compile(r'\byear\s*=\s*(?:\{+|"|\b)\s*(\d+)[\}"]?,?', re.IGNORECASE)
 
-if sublime.version() < '3000':
-    def _get_people_long(people):
-        return u' and '.join([unicode(x) for x in people])
-else:
-    def _get_people_long(people):
-        return u' and '.join([str(x) for x in people])
+# This may speed things up
+# So far this captures: the tag, and the THREE possible groups
+multip = re.compile(r'\b(author|title|year|editor|journal|eprint)\s*=\s*(?:\{|"|\b)(.+?)(?:\}+|"|\b)\s*,?\s*\Z',re.IGNORECASE)
 
-def _get_people_short(people):
-    if len(people) <= 2:
-        return u' & '.join([x.last for x in people])
-    else:
-        return people[0].last + u', et al.'
-
-def remove_latex_commands(s):
-    u'''
-    Simple function to remove any LaTeX commands or brackets from the string,
-    replacing it with its contents.
-    '''
-    chars = []
-    FOUND_SLASH = False
-
-    for c in s:
-        if c == '{':
-            # i.e., we are entering the contents of the command
-            if FOUND_SLASH:
-                FOUND_SLASH = False
-        elif c == '}':
-            pass
-        elif c == '\\':
-            FOUND_SLASH = True
-        elif not FOUND_SLASH:
-            chars.append(c)
-        elif c.isspace():
-            FOUND_SLASH = False
-
-    return ''.join(chars)
-
-# wrapper to implement a dict-like interface for bibliographic entries
-# returning formatted value, if it is available
-class EntryWrapper(Mapping):
-    def __init__(self, entry):
-        self.entry = entry
-
-    def __getitem__(self, key):
-        if not key:
-            return u''
-
-        key = key.lower()
-        result = None
-
-        short = False
-        if key.endswith('_short'):
-            short = True
-            key = key[:-6]
-
-        if key == 'keyword' or key == 'citekey':
-            return self.entry.cite_key
-
-        if key in Name.NAME_FIELDS:
-            people = []
-            for x in tokenize_list(self.entry[key]):
-                if x.strip() == '':
-                    continue
-
-                try:
-                    people.append(Name(x))
-                except:
-                    print(u'Error handling field "{0}" with value "{1}"'.format(
-                        key, x
-                    ))
-                    traceback.print_exc()
-
-            if len(people) == 0:
-                return u''
-
-            if short:
-                result = _get_people_short(people)
-            else:
-                result = _get_people_long(people)
-
-        if not result:
-            result = self.entry[key]
-
-        return remove_latex_commands(codecs.decode(result, 'latex'))
-
-    def __iter__(self):
-        return iter(self.entry)
-
-    def __len__(self):
-        return len(self.entry)
 
 class TraditionalBibliographyPlugin(LaTeXToolsPlugin):
     def get_entries(self, *bib_files):
         entries = []
-        parser = Parser()
         for bibfname in bib_files:
-            cache_name = "bib_" + hashlib.md5(bibfname.encode("utf8")).hexdigest()
+            cache_name = "tradbib_" + hashlib.md5(bibfname.encode("utf8")).hexdigest()
             try:
                 modified_time = os.path.getmtime(bibfname)
 
@@ -131,15 +54,56 @@ class TraditionalBibliographyPlugin(LaTeXToolsPlugin):
                 sublime.status_message("Cannot open bibliography file %s !" % (bibfname,))
                 continue
             else:
-                bib_data = parser.parse(bibf.read())
+                bib_data = bibf.readlines()
 
-                print ('Loaded %d bibitems' % (len(bib_data)))
-
-                for key in bib_data:
-                    entry = bib_data[key]
-                    if entry.entry_type in ('xdata', 'comment', 'string'):
+                entry = {}
+                for line in bib_data:
+                    line = line.strip()
+                    # Let's get rid of irrelevant lines first
+                    if line == "" or line[0] == '%':
                         continue
-                    entries.append(EntryWrapper(entry))
+                    if line.lower()[0:8] == "@comment":
+                        continue
+                    if line.lower()[0:7] == "@string":
+                        continue
+                    if line.lower()[0:9] == "@preamble":
+                        continue
+                    if line[0] == "@":
+                        if 'keyword' in entry:
+                            entries.append(entry)
+                            entry = {}
+
+                        kp_match = kp.search(line)
+                        if kp_match:
+                            entry['keyword'] = kp_match.group(1)
+                        else:
+                            print("Cannot process this @ line: " + line)
+                            print(
+                                "Previous keyword (if any): " +
+                                entry.get('keyword', '')
+                            )
+                        continue
+
+                    # Now test for title, author, etc.
+                    # Note: we capture only the first line, but that's OK for our purposes
+                    multip_match = multip.search(line)
+                    if multip_match:
+                        key = multip_match.group(1).lower()
+                        value = multip_match.group(2)
+
+                        if key == 'title':
+                            value = value.replace(
+                                '{\\textquoteright}', ''
+                            ).replace('{', '').replace('}', '')
+                        entry[key] = value
+                    continue
+
+                # at the end, we have a single record
+                if 'keyword' in entry:
+                    entries.append(entry)
+
+
+                print ('Loaded %d bibitems' % (len(entries)))
 
                 try:
                     cache.write_global(cache_name, (modified_time, entries))
@@ -156,6 +120,3 @@ class TraditionalBibliographyPlugin(LaTeXToolsPlugin):
 
             print("Found %d total bib entries" % (len(entries),))
         return entries
-
-    def on_insert_citation(self, keyword):
-        print('Inserted {0}'.format(keyword))
